requirements.txt

Django==2.2.5
django-filter==2.2.0
djangorestframework==3.11.0
mysqlclient == 1.3.13 
django-mysql == 2.2.0 
